rename Settings class
