+++ data/settings/src/main/java/com/google/jetpackcamera/settings/model/CameraAppSettings.kt
@@ -19,15 +19,15 @@ package com.google.jetpackcamera.settings.model
 /**
  * Data layer representation for settings.
  */
-data class Settings(
+data class CameraAppSettings(
     val default_front_camera : Boolean,
     val dark_mode_status : DarkModeStatus,
     val field : String = "todo"
 )
 
 
-fun getDefaultSettings(): Settings{
-    return Settings(
+fun getDefaultSettings(): CameraAppSettings{
+    return CameraAppSettings(
         default_front_camera = false,
         dark_mode_status = DarkModeStatus.SYSTEM
     )
